Move alternative symbol resolution from resolution.rs to symbol_db.rs
This seems like a better place for this code and allows some of the fields on
SymbolDb to be made private.
